fix(lib): tighten crate public API -- demote internal tool helpers to pub(crate)
