+++ crates/aptu-coder/src/tests.rs
@@ -1,5 +1,6 @@
 #[cfg(test)]
 use super::*;
+use crate::tools::common::summary_cursor_conflict;
 use crate::tools::exec_command::{build_exec_command, handle_output_persist, strip_cd_prefix};
 use crate::validation::validate_path_in_dir;
 use aptu_coder_core::traversal;
